test(app): split pawwork sidebar source tests
